+++ sqlite3/lib/src/hook/compile/description.dart
@@ -7,9 +7,9 @@ import 'package:crypto/crypto.dart';
 import 'package:hooks/hooks.dart';
 import 'package:path/path.dart' as p;
 
-import 'assets.dart';
-import 'asset_hashes.dart';
-import 'utils.dart';
+import '../assets.dart';
+import '../asset_hashes.dart';
+import '../utils.dart';
 
 /// Possible sources to obtain a `libsqlite3.so` (or the equivalent for other
 /// platforms).
@@ -495,4 +495,5 @@ const _defaultDefines = '''
   SQLITE_HAVE_LOCALTIME_S
   SQLITE_HAVE_MALLOC_USABLE_SIZE
   SQLITE_HAVE_STRCHRNUL
+  SQLITE_ENABLE_BATCH_ATOMIC_WRITE
 ''';