+++ sqlite3/lib/src/wasm/injected_values.dart
@@ -263,6 +263,15 @@ final class DartBridgeCallbacks {
     return file.toDartObject.xDeviceCharacteristics;
   }
 
+  @JSExport()
+  int xFileControl(
+    ExternalDartReference<VirtualFileSystemFile> file,
+    int op,
+    int pArg,
+  ) {
+    return file.toDartObject.xFileControl(SqliteFileControl(op), pArg);
+  }
+
   @JSExport('dispatch_()v')
   void dispatchVoid(ExternalDartReference<void Function()> fn) {
     fn.toDartObject();