+++ sqlite3_wasm_build/src/bridge.h
@@ -46,6 +46,8 @@ import_dart("xLock") extern int xLock(__externref_t, int);
 import_dart("xUnlock") extern int xUnlock(__externref_t, int);
 import_dart("xCheckReservedLock") extern int xCheckReservedLock(__externref_t,
                                                                 int* pResOut);
+import_dart("xFileControl") extern int xFileControl(__externref_t, int op,
+                                                    void* pArg);
 import_dart("xDeviceCharacteristics") extern int xDeviceCharacteristics(
     __externref_t);